% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{createLink}
\alias{createLink}
\title{createLink function}
\usage{
createLink(val, on_click_js)
}
\arguments{
\item{val}{string}

\item{on_click_js}{string
<a class="btn btn-primary">Info</a>'}
}
\description{
TODO: Change documentation of this function
}
