% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripr-package.R
\docType{package}
\name{tripr}
\alias{tripr-package}
\alias{tripr}
\title{tripr}
\description{
T-cell Receptor/Immunoglobulin Profiler (TRIP)
}
\details{
The only function you're likely to need from tripr is [run_app()].
Otherwise refer to the vignettes for using tripr.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BiodataAnalysisGroup/tripr}
  \item Report bugs at \url{https://github.com/BiodataAnalysisGroup/tripr/issues}
}

}
\author{
\strong{Maintainer}: Nikolaos Pechlivanis \email{inab.bioinformatics@lists.certh.gr}

Authors:
\itemize{
  \item Maria Th. Kotouza
  \item Katerina Gemenetzi
  \item Chrysi Galigalidou
  \item Elisavet Vlachonikola
  \item Andreas Agathangelidis
  \item Raphael Sandaltzopoulos
  \item Pericles A. Mitkas
  \item Kostas Stamatopoulos
  \item Anastasia Chatzidimitriou
  \item Fotis E. Psomopoulos
  \item Iason Ofeidis
  \item Aspasia Orfanou
}

}
