% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripr-package.R
\docType{package}
\name{tripr}
\alias{tripr}
\title{tripr}
\description{
T-cell Receptor/Immunoglobulin Profiler (TRIP)
}
\details{
The only function you're likely to need from tripr is [run_app()].
Otherwise refer to the vignettes for using tripr.
}
